requirements.txt
# requirements.txt
streamlit>=1.32.0
google-generativeai>=0.3.0
serpapi>=0.1.5
requests>=2.31.0